permissions for all entities

+++ app/Http/Controllers/TestCaseController.php
@@ -14,6 +14,10 @@ class TestCaseController extends Controller
 {
     public function store(Request $request)
     {
+        if(!auth()->user()->can('add_edit_test_cases')) {
+            abort(403);
+        }
+
         $testCase = new TestCase();
 
         $testCase->title = $request->title;
@@ -36,7 +40,12 @@ class TestCaseController extends Controller
         ];
     }
 
-    public function update(Request $request) {
+    public function update(Request $request)
+    {
+        if(!auth()->user()->can('add_edit_test_cases')) {
+            abort(403);
+        }
+
         $testCase = TestCase::findOrFail($request->id);
 
         $testCase->title = $request->title;
@@ -59,7 +68,12 @@ class TestCaseController extends Controller
         ];
     }
 
-    public function destroy(Request $request) {
+    public function destroy(Request $request)
+    {
+        if(!auth()->user()->can('delete_test_cases')) {
+            abort(403);
+        }
+
         $testCase = TestCase::findOrFail($request->id);
         $testCase->delete();
     }
