+++ app/Http/Controllers/TestRunController.php
@@ -21,6 +21,7 @@ class TestRunController extends Controller
         $results[$request->test_case_id] = $request->status;
         $testRun->saveResults($results);
     }
+
     /*****************************************
      *  PAGES
      *****************************************/
@@ -61,9 +62,12 @@ class TestRunController extends Controller
             ->with('results', $results);
     }
 
-
     public function create($project_id)
     {
+        if(!auth()->user()->can('add_edit_test_runs')) {
+            abort(403);
+        }
+
         $project = Project::findOrFail($project_id);
         $testPlans = TestPlan::all();
 
@@ -74,6 +78,10 @@ class TestRunController extends Controller
 
     public function edit($project_id, $test_run_id)
     {
+        if(!auth()->user()->can('add_edit_test_runs')) {
+            abort(403);
+        }
+
         $project = Project::findOrFail($project_id);
         $testRun = TestRun::findOrFail($test_run_id);
 
@@ -89,6 +97,10 @@ class TestRunController extends Controller
 
     public function store(Request $request)
     {
+        if(!auth()->user()->can('add_edit_test_runs')) {
+            abort(403);
+        }
+
         $request->validate([
             'title' => 'required',
             'test_plan_id' => 'required',
@@ -106,6 +118,10 @@ class TestRunController extends Controller
 
     public function update(Request $request)
     {
+        if(!auth()->user()->can('add_edit_test_runs')) {
+            abort(403);
+        }
+
         $testRun = TestRun::findOrFail($request->id);
 
         $testRun->title = $request->title;
@@ -116,6 +132,10 @@ class TestRunController extends Controller
 
     public function destroy(Request $request)
     {
+        if(!auth()->user()->can('delete_test_runs')) {
+            abort(403);
+        }
+
         $testRun = TestRun::findOrFail($request->id);
         $testRun->delete();
         return redirect()->route('test_run_list_page', $request->project_id);
